third_party/WebKit/LayoutTests/http/tests/security/media-element-audio-source-node-cross-origin-expected.txt

CONSOLE WARNING: line 21: The Web Audio autoplay policy will be re-enabled in Chrome 70 (October 2018). Please check that your website is compatible with it. https://goo.gl/7K7WLu
CONSOLE MESSAGE: MediaElementAudioSource outputs zeroes due to CORS access restrictions for http://localhost:8000/security/resources/webaudio/laughter.wav
MediaElementAudioSourceNode with cross-origin source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS All samples correctly zeroed.
PASS successfullyParsed is true

TEST COMPLETE
